worker/requirements.txt
apscheduler>=3.10.4
asyncpg>=0.29.0
livekit-agents[silero,turn-detector,deepgram,google]~=1.4
livekit-plugins-noise-cancellation~=0.2
pydantic>=2.6.0
python-dotenv>=1.0.0
